# openapi.yaml
openapi: 3.0.3

info:
  title: TV Tracker API
  description: API for searching and retrieving movies and TV shows from TMDB
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Media
    description: Search and retrieve media
  - name: Users
    description: User management
  - name: Lists
    description: User lists and tracking

paths:

  /media/search:
    get:
      tags:
        - Media
      summary: Search media
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [movie, tv]
      responses:
        '200':
          description: Successful search
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /media/{type}/{id}:
    get:
      tags:
        - Media
      summary: Get media details
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [movie, tv]
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Media details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaDetails'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /media/tv/{seriesId}/season/{seasonNum}:
    get:
      tags:
        - Media
      summary: Get TV season details
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
        - name: seasonNum
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Season details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonDetails'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:

  schemas:

    Media:
      type: object
      required:
        - id
        - title
        - type
      properties:
        id:
          type: integer
          example: 550
        title:
          $ref: '#/components/schemas/Title'
        imageUrl:
          type: string
        type:
          type: string
          enum: [movie, tv]
        releaseYear:
          type: integer
          format: int32
        score:
          type: number
          format: float

    MediaDetails:
      type: object
      required:
        - id
        - title
        - type
      properties:
        id:
          type: integer
        title:
          $ref: '#/components/schemas/Title'
        imageUrl:
          type: string
        type:
          type: string
          enum: [movie, tv]
        releaseYear:
          type: integer
          format: int32
        endYear:
          type: integer
          format: int32
        score:
          type: number
          format: float
        description:
          type: string
        duration:
          type: integer
          example: 120
          description: Duration in minutes (movie runtime or average episode duration for series)
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/Season'
        totalEpisodes:
          type: integer
        relatedMedia:
          type: array
          items:
            $ref: '#/components/schemas/RelatedMedia'
        status:
          type: string

    Season:
      type: object
      properties:
        id:
          type: integer
        totalEpisodes:
          type: integer
        imageUrl:
          type: string
        releaseYear:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        score:
          type: number
          format: float
        seasonNumber:
          type: integer

    SeasonDetails:
      type: object
      properties:
        id:
          type: integer
        totalEpisodes:
          type: integer
        imageUrl:
          type: string
        releaseYear:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        score:
          type: number
          format: float
        seasonNumber:
          type: integer
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/Episode'

    Episode:
      type: object
      properties:
        id:
          type: integer
        episodeNumber:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        score:
          type: number
          format: float
        description:
          type: string
        releaseDate:
          type: string
          format: date
        duration:
          type: integer

    RelatedMedia:
      type: object
      required:
        - relatedType
        - media
      properties:
        relatedType:
          type: string
          enum: [sequel, prequel, alternative]
        media:
          $ref: '#/components/schemas/Media'

    Title:
      type: object
      required:
        - english
      properties:
        english:
          type: string
        original:
          type: string

    ErrorResponse:
      type: object
      required:
        - message
        - status
      properties:
        message:
          type: string
        status:
          type: integer