# src/main/java/com/lbranco/tv_tracker_api/media/model/Title.java
package com.lbranco.tv_tracker_api.media.model;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class Title {
    private String english;
    private String original;
}